Make meta fields optional:
You can skip defining the fields we should extract. In that case, we
will use all the fields with the defined prefix.

We will generate a unique prefix if none is provided, although we strongly
urge you to always define the prefix.

If no pk is given, we'll fallback to all fields as usual.

Signed-off-by: Pedro Melo <melo@simplicidade.org>