Add support for into meta key:
Allows us to move the contents of a specific meta block into a HashRef.
It is usefull if you don't want to mix your fields with nested keys,
thus moving all fields into a hashRef under the `into` value as key.

If the output without into is:

~~~perl
{
  k => 1,
  f1 => 'field 1',
  f2 => 'field 2',
  n1 => [
    ... nested1 ...
  ],
  n2 => [
    ... nested2 ...
  ],
}
~~~

Adding a `into => 'top'` to the that meta block definition would
transform it into:

~~~perl
{
  top => {
    k => 1,
    f1 => 'field 1',
    f2 => 'field 2',
  },
  n1 => [
    ... nested1 ...
  ],
  n2 => [
    ... nested2 ...
  ],
}
~~~

Signed-off-by: Pedro Melo <melo@simplicidade.org>